# perf(s3stream): speedup wal recovery