--- a/s3stream/src/main/java/com/automq/stream/s3/wal/util/WALBlockDeviceChannel.java
+++ b/s3stream/src/main/java/com/automq/stream/s3/wal/util/WALBlockDeviceChannel.java
@@ -295,9 +295,10 @@ public class WALBlockDeviceChannel implements WALChannel {
     }
 
     @Override
-    public int read(ByteBuf dst, long position) throws IOException {
+    public int read(ByteBuf dst, long position, int length) throws IOException {
         long start = position;
-        long end = position + dst.writableBytes();
+        length = Math.min(length, dst.writableBytes());
+        long end = position + length;
         long alignedStart = WALUtil.alignSmallByBlockSize(start);
         long alignedEnd = WALUtil.alignLargeByBlockSize(end);
         int alignedSize = (int) (alignedEnd - alignedStart);