--- a/s3stream/src/main/java/com/automq/stream/s3/wal/util/WALCachedChannel.java
+++ b/s3stream/src/main/java/com/automq/stream/s3/wal/util/WALCachedChannel.java
@@ -17,6 +17,7 @@
 
 package com.automq.stream.s3.wal.util;
 
+import com.automq.stream.s3.DirectByteBufAlloc;
 import io.netty.buffer.ByteBuf;
 import java.io.IOException;
 
@@ -28,11 +29,14 @@ public class WALCachedChannel implements WALChannel {
     private static final int DEFAULT_CACHE_SIZE = 1 << 20;
 
     private final WALChannel channel;
-    private final int cacheSize;
+    private final int cacheSizeWant;
 
-    private WALCachedChannel(WALChannel channel, int cacheSize) {
+    private ByteBuf cache;
+    private long cachePosition = -1;
+
+    private WALCachedChannel(WALChannel channel, int cacheSizeWant) {
         this.channel = channel;
-        this.cacheSize = cacheSize;
+        this.cacheSizeWant = cacheSizeWant;
     }
 
     public static WALCachedChannel of(WALChannel channel) {
@@ -43,21 +47,66 @@ public class WALCachedChannel implements WALChannel {
         return new WALCachedChannel(channel, cacheSize);
     }
 
+    /**
+     * As we use a common cache for all threads, we need to synchronize the read.
+     */
     @Override
-    public int read(ByteBuf dst, long position) throws IOException {
-        // TODO
-        return this.channel.read(dst, position);
+    public synchronized int read(ByteBuf dst, long position) throws IOException {
+        long start = position;
+        int length = dst.writableBytes();
+        long end = position + length;
+        ByteBuf cache = getCache();
+        boolean fallWithinCache = cachePosition >= 0 && cachePosition <= start && end <= cachePosition + cache.readableBytes();
+        if (!fallWithinCache) {
+            cache.clear();
+            cachePosition = start;
+            if (cachePosition + cache.writableBytes() > channel.capacity()) {
+                // The cache is larger than the channel capacity.
+                cachePosition = channel.capacity() - cache.writableBytes();
+                assert cachePosition >= 0;
+            }
+            channel.read(cache, cachePosition);
+        }
+        // Now the cache is ready.
+        int relativePosition = (int) (start - cachePosition);
+        dst.writeBytes(cache, relativePosition, length);
+        return length;
+    }
+
+    @Override
+    public void close() {
+        releaseCache();
+        this.channel.close();
+    }
+
+    /**
+     * Release the cache if it is not null.
+     * This method should be called when no more {@link #read}s will be called to release the allocated memory.
+     */
+    public synchronized void releaseCache() {
+        if (this.cache != null) {
+            this.cache.release();
+            this.cache = null;
+        }
+        this.cachePosition = -1;
+    }
+
+    /**
+     * Get the cache. If the cache is not initialized, initialize it.
+     * Should be called under synchronized.
+     */
+    private ByteBuf getCache() {
+        if (this.cache == null) {
+            // Make sure the cache size is not larger than the channel capacity.
+            int cacheSize = (int) Math.min(this.cacheSizeWant, this.channel.capacity());
+            this.cache = DirectByteBufAlloc.byteBuffer(cacheSize);
+        }
+        return this.cache;
     }
 
     @Override
     public void open(CapacityReader reader) throws IOException {
         this.channel.open(reader);
-    }
-
-    @Override
-    public void close() {
-        // TODO: maybe release cache
-        this.channel.close();
     }
 
     @Override