resolve https://github.com/AutoMQ/automq-for-kafka/issues/573

On AWS EBS gp3 (3000 IOPS, 125 MB/s):
Recover 65536 messages (16 KiB each, 1 GiB in total) cost 92.7s -> 15.3s
Recover 65536 messages (4 KiB each, 256 MiB in total) cost 65.8s -> 2.5s